Convert all PostgreSQL integer keys to bigint
There is one special exception in here: the unread_user_ids column in
the message_conversations table had to be left as an integer array,
since the PostgreSQL intarray extension doesn't work with bigints. The
trigger that updates that column also needed a minor tweak.

This isn't good, but I don't really like how that was done anyway (it
was for the purpose of group messages that don't even exist), so it
could probably just be eliminated.

+++ tildes/sql/init/triggers/message_replies/message_conversations.sql
@@ -11,8 +11,8 @@ BEGIN
             SET num_replies = num_replies + 1,
                 last_reply_time = NEW.created_time,
                 unread_user_ids = unread_user_ids |
-                    CASE WHEN sender_id = NEW.sender_id THEN recipient_id
-                        ELSE sender_id
+                    CASE WHEN sender_id = NEW.sender_id THEN recipient_id::integer
+                        ELSE sender_id::integer
                     END
             WHERE conversation_id = NEW.conversation_id;
     ELSIF (TG_OP = 'DELETE') THEN
